feat: swizzable MarkdownActionsDropdown component
