+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "docusaurus-markdown-source-plugin",
-  "version": "2.2.5",
+  "version": "2.3.0",
   "description": "A lightweight Docusaurus plugin that exposes your markdown files as raw .md URLs, perfect for LLMs and documentation tools",
   "main": "index.js",
   "scripts": {
